Module Refactor
* Moves modules from SoftLayer/CLI/*.py to SoftLayer/CLI/modules/*.py
* Moves code from SoftLayer/CLI/__init__.py to SoftLayer/CLI/core.py,
  SoftLayer/CLI/environment.py, SoftLayer/CLI/helpers.py
