+++ SoftLayer/CLI/modules/cci.py
