+++ SoftLayer/CLI/modules/image.py
